requirements.txt
streamlit
requests
tweepy
praw
nltk
textblob
pandas
plotly
pymongo
auth0-python
python-dotenv
schedule
